requirements.txt
crewai>=0.70.0
crewai-tools>=0.69.0
arxiv>=2.1.0
python-dotenv>=1.0.0